+++ Content.Shared/Bed/Components/HealOnBuckleHealing.cs
